renaming plain-text file "VERISON" to "VERSION.py" and explicitly importing in `__version__.py` so pyinstaller will automatically detect VERSION as a dependency.
The original method using a plain text file results in a run-time error when pyinstaller is used to generate a single file executable with the following line:

`pyinstaller -F -w <hart_protocol_user_script.py>`

where `<hart_protocol_user_script.py>` is a script that imports `hart_protocol`

pyinstaller was not detecting file *VERSION* as a required dependency and it wasn't being included in the executable image.  There may also be another method of resolving this issue on the pyinstaller side by using --add-data or possibly a hook file (my first attempts were not successful).